fix: remove support for prop called `render`. Only support `children`
This is done to bring downshift into conformance with the community
convention of using the children prop (for example, the React Team chose
`children` with the official context API). We'd prefer to reduce
cognitive overhead by not having two ways to do the same thing.

Closes #433

BREAKING CHANGE: If you use the `render` prop, rename it to `children`.