feat: add ARIA specification of First and Last option highlight (#590)
**What**:
Added highlight on first and last options when user opens the dropdown via Up/Down Arrows.

**Why**:
According to ARIA, if dropdown is collapsed and focus is on the text input:
- Arrow Down should open menu **and** highlight on option.
- Arrow Up should open menu **and** highlight last option.

**How**:

Modified `moveHighlightedIndex` in dropdown.js, specifically the `else` branch. There I open the menu, and then check if the opening action was made by Arrow up or down. I also check if there are any items in the dropdown.

If yes, I call `setHighlightedIndex` with either `itemCount - 1` or `0`, depending on the value of the event type (arrow up or arrow down), in order to comply to ARIA specifications.

I also modified some unit tests and end-to-end tests in order to reflect the changes. Also added 2 more end-to-end tests that test only this new behaviour.

Closes https://github.com/paypal/downshift/issues/589