feat(initial*): add `initial*` props for all state
Before the `default*` props sort of served two purposes (the initial
value and the value to reset certain state to). Now we've actually
removed the `defaultSelectedItem` and `defaultInputValue` because we
never reset the values to the initial value for those. And we introduce
new `initial*` props for all state to take over what the default ones
were used for in the constructor.

Closes: #467

BREAKING CHANGE: If you're using any of the props that start with `default`, you probably want to change that to `initial` instead. `defaultSelectedItem` and `defaultInputValue` were also removed in favor of `initialSelectedItem and `initialInputValue` (respectively).