docs: Add null-check note for itemToString (#975)
This commit includes one small, overall style change, as well:
most U.S. style guides recommend that "e.g." be followed by a comma,[0]
whereas the downshift docs leave it bare.

The core changes in this commit use the phrase "e.g." in a few places,
so in the interest of consistency, this commit applies the comma both
in new additions and existing instances of "e.g." throughout the docs.

[0]: https://www.dailywritingtips.com/comma-after-i-e-and-e-g/