Fix typings
The `type` field [is set on the state](https://github.com/downshift-js/downshift/blob/1c7698080805e05f5180dd223cbc17932a85fd42/src/downshift.js#L351) that is passed to the `onInputValueChange` callback prop but the typings don't include it. This field is set correctly for other state change callbacks, it is just missing for this one. Let's fix it.