fix: update element ids when related props change
We used to memoize the element id props on the first render and then use
those throughout the lifecycle of the component. This could cause ids in
the markup to be out of sync from the ids provided via props. This
change updates the internal element id getter references whenever any id
related props change to propagate those props changes to the markup.